requirements.txt
requests[security]
crowdstrike-falconpy==1.2.13
humiolib==0.2.4
python-dateutil==2.8.2
